+++ src/endstone/core/entity/components/flag_components.h
@@ -17,6 +17,9 @@
 namespace endstone::core {
 struct InternalDisconnectFlagComponent {};
 struct InternalRemoveFlagComponent {};
+struct InternalSignInteractFlagComponent {};
+struct InternalSignOpenFlagComponent {};
+struct InternalSignPlaceFlagComponent {};
 struct InternalSpawnChangeFlagComponent {};
 struct InternalTeleportFlagComponent {};
 struct MobHurtFlagComponent {};